chore: move vercel api to root-level
