+++ vercel.json
@@ -2,14 +2,8 @@
   "buildCommand": "npm run build",
   "outputDirectory": "dist",
   "functions": {
-    "src/vercel/api/*.ts": {
+    "api/*.ts": {
       "runtime": "nodejs18.x"
     }
-  },
-  "rewrites": [
-    {
-      "source": "/api/(.*)",
-      "destination": "/src/vercel/api/$1"
-    }
-  ]
+  }
 }
